# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Roil Finance API
  description: |
    Backend API for Roil Finance — an institutional-grade portfolio rebalancing
    platform built on Canton Network with Daml smart contracts.

    Supports portfolio management, DCA scheduling, reward tracking, market data,
    token transfers, yield compounding, and platform administration.
  version: 1.0.0
  contact:
    name: Roil Finance
    url: https://roil.app
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.roil.app
    description: Production

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing `actAs` array of Canton party identifiers.
        Algorithms supported: HS256 (dev), RS256, ES256.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Something went wrong"
    AssetId:
      type: object
      required: [symbol, admin]
      properties:
        symbol:
          type: string
          example: "CBTC"
        admin:
          type: string
          example: "admin-party::1220abc"
    TargetAllocation:
      type: object
      required: [asset, targetPct]
      properties:
        asset:
          $ref: '#/components/schemas/AssetId'
        targetPct:
          type: number
          minimum: 0
          maximum: 100
          example: 25
    TriggerMode:
      oneOf:
        - type: string
          enum: [Manual]
        - type: object
          properties:
            tag:
              type: string
              enum: [DriftThreshold]
            value:
              type: number
              minimum: 0.1
              maximum: 100
              example: 5.0
        - type: object
          properties:
            tag:
              type: string
              enum: [PriceCondition]
            value:
              type: object
              properties:
                conditionAsset:
                  type: string
                targetPrice:
                  type: number
                conditionAction:
                  type: string
                  enum: [sell_above, buy_below]
...
paths:
...
  /api/portfolio:
    post:
      tags: [Portfolio]
      summary: Create a new portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user, targets, triggerMode]
              properties:
                user:
                  type: string
                targets:
                  type: array
                  items:
                    $ref: '#/components/schemas/TargetAllocation'
                  minItems: 2
                triggerMode:
                  $ref: '#/components/schemas/TriggerMode'
      responses:
        '201':
          description: Portfolio created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Validation error or targets do not sum to 100%
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/dca:
    post:
      tags: [DCA]
      summary: Create a new DCA schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user, sourceAsset, targetAsset, amountPerBuy, frequency]
              properties:
                user:
                  type: string
                sourceAsset:
                  $ref: '#/components/schemas/AssetId'
                targetAsset:
                  $ref: '#/components/schemas/AssetId'
                amountPerBuy:
                  type: number
                frequency:
                  type: string
                  enum: [Hourly, Daily, Weekly, Monthly]
      responses:
        '201':
          description: DCA schedule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...